+++ tools/Polygraphy/polygraphy/__init__.py
@@ -1,3 +1,3 @@
 import polygraphy.config
 
-__version__ = "0.44.2"
+__version__ = "0.45.0"