+++ tools/Polygraphy/polygraphy/tools/inspect/subtool/data.py
@@ -14,13 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from polygraphy import util
+import sys
+
+from polygraphy import mod, util
 from polygraphy.common import TensorMetadata
 from polygraphy.comparator import RunResults
 from polygraphy.comparator import util as comp_util
 from polygraphy.json import load_json
 from polygraphy.logger import G_LOGGER
 from polygraphy.tools.base import Tool
+
+np = mod.lazy_import("numpy")
 
 
 class Data(Tool):
@@ -44,8 +48,30 @@ class Data(Tool):
             "-s", "--show-values", help="Show values of the tensors instead of just metadata", action="store_true"
         )
         parser.add_argument("--histogram", help="Show a histogram of the value distribution", action="store_true")
+        parser.add_argument(
+            "-n",
+            "--num-items",
+            help="The number of values to show at the beginning and end of each dimension when printing arrays. "
+            "Use a value of -1 to show all elements in the array. "
+            "Defaults to 3.",
+            type=int,
+            default=3,
+        )
+        parser.add_argument(
+            "--line-width",
+            help="The number of characters per line when displaying arrays. "
+            "Use a value of -1 to insert line breaks only at dimension end points. "
+            "Defaults to 75.",
+            type=int,
+            default=75,
+        )
 
     def run_impl(self, args):
+        np.set_printoptions(
+            edgeitems=sys.maxsize if args.num_items == -1 else args.num_items,
+            linewidth=sys.maxsize if args.line_width == -1 else args.line_width,
+        )
+
         # Note: It's important we have encode/decode JSON methods registered
         # for the types we care about, e.g. RunResults. Importing the class should generally guarantee this.
         data = load_json(args.path)